pyproject.toml
[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
  /(
    \.(git|hg|mypy_cache|tox|venv)
    | _?build
    | dist
  )/
  '''

[tool.commitizen]
name = "cz_customize"
version_files = ["pyproject.toml:version"]

[tool.commitizen.customize]
schema_pattern = "(init|build|ci|docs|feat|fix|perf|refactor|style|test|chore|revert|bump)(\\(\\S+\\))?!?:(\\s.*)"

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 79
lines_after_imports = 2
multi_line_output = 3
use_parentheses = true

[tool.poetry]
name = "wraeblast"
version = "0.1.0-alpha.0"
description = "Tools for Path of Exile filter generation and economy data analysis."
authors = ["David Gidwani <david.gidwani@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
wraeblast = "wraeblast.cmd:main"

[tool.poetry.dependencies]
python = "^3.9"
lark = "^0.11.3"
matplotlib = "^3.4.2"
pydantic = "^1.8.2"
cleo = "^0.8.1"
structlog = "^21.1.0"
requests = "^2.25.1"
uplink = { extras = ["aiohttp"], version = "^0.9.4" }
colour = "^0.1.5"
diskcache = "^5.2.1"
python-json-logger = "^2.0.1"
glom = "^20.11.0"
"ruamel.yaml" = "^0.17.5"
Jinja2 = "^3.0.1"
colorcet = "^2.0.6"
palettable = "^3.3.0"
mergedeep = "^1.3.4"
pydub = "^0.25.1"
boto3 = "^1.18.26"
pandas = "^1.3.2"
pyttsx3 = "^2.90"
colormath = "^3.0.0"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
isort = "^5.8.0"
pre-commit = "^2.12.1"
ipython = "^7.23.1"
colorama = "^0.4.4"
pytest = "^6.2.4"
flameprof = "^0.4"
plotly = "^5.1.0"
mkdocs-material = "^7.1.11"
mkdocstrings = "^0.15.2"
commitizen = "^2.17.13"
pymdown-extensions = "^8.2"
mkdocs-literate-nav = "^0.4.0"
mkdocs-minify-plugin = "^0.4.0"
tox = "^3.24.3"

[tool.pytest.ini_options]
filterwarnings = ['ignore::DeprecationWarning']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
